Refactor: rename files from index.tsx to component_name.tsx
